Dev (#2)
* Fix DVC configuration to correct S3 bucket path

This commit addresses an issue in the DVC configuration by updating the S3 bucket path to ensure proper data storage and retrieval. The previous path was incorrect, which could lead to data management errors. This change is crucial for maintaining the integrity of data versioning and ensuring that the project functions as intended.

* Add DVC configuration for data versioning

* Update .gitignore to exclude model files and improve project cleanliness

This commit modifies the .gitignore file to include the exclusion of the 'models' directory, ensuring that model files are not tracked by version control. Additionally, it maintains the exclusion of '.pth' files. These changes enhance project organization by preventing unnecessary files from cluttering the repository, aligning with best practices for version control.

* Change file permissions for setup-repo-connection.sh to make it executable

This commit updates the file mode of setup-repo-connection.sh from 644 to 755, allowing it to be executed as a script. This change is necessary for ensuring that the script can be run directly in the terminal, facilitating easier setup of the repository connection. Proper file permissions are essential for script functionality and user convenience.

* Update Argo Workflows and AWS credentials management

- Enhance README.md with AWS credentials setup instructions.
- Modify ml-pipeline.yaml to use a new S3 bucket and include data versioning parameters.
- Update deploy-s3-bucket.sh to generate unique bucket names with timestamps.
- Add AWS credentials management in run-argo-pipeline.sh to create/update secrets.
- Refactor update-dvc-config.sh and update-aws-credentials.sh for improved paths and functionality.
- Introduce a new .dvc/config file for DVC S3 configuration.

---------

Co-authored-by: Andrii Maslovskyi <amaslovs@cisco.com>