Remove obsolete Argo workflow and deployment script files
This commit deletes the `git-repo-workflow.yaml`, `ml-pipeline.yaml`, and `deploy-s3-bucket.sh` files from the repository. These files are no longer needed as the project has transitioned to a new workflow structure and deployment strategy. Removing these obsolete files helps maintain a clean and organized codebase, ensuring that only relevant and up-to-date resources are available for development and deployment.